like and share fix
I've fixed the codes for like and share. i've tested already. both like
and unlike works fine. just the loading a bit slow. need to fix the
removal of the add comment box, and the share. The share not dynamic. i
can't do that now, need help with the jquery codes. but confirm share
adds the count and points.